Move estimate-fee
